pyproject.toml
[project]
name = "aiogram-bot-template"
version = "0.1.0"
description = ""
authors = [
    {name = "niko0701",email = "totsuka.totoro07012005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
